backend/requirements/base.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-countries==6.1.2
django-simple-history==2.11.0
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
